testing/web-platform/tests/web-animations/timing-model/timelines/timelines.html

<!doctype html>
<meta charset=utf-8>
<title>Timelines</title>
<link rel="help" href="https://w3c.github.io/web-animations/#timelines">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<div id="log"></div>
<script>
'use strict';

promise_test(function(t) {
  const valueAtStart = document.timeline.currentTime;
  const timeAtStart = window.performance.now();
  while (window.performance.now() - timeAtStart < 50) {
    // Wait 50ms
  }
  assert_equals(document.timeline.currentTime, valueAtStart,
    'document.timeline.currentTime does not change within a script block');
  return waitForAnimationFrames(1).then(function() {
    assert_greater_than(document.timeline.currentTime, valueAtStart,
      'document.timeline.currentTime increases between script blocks');
  });
}, 'document.timeline.currentTime liveness tests');

async_test(function(t) {
  const iframe = document.createElement('iframe');
  iframe.width = 10;
  iframe.height = 10;

  iframe.addEventListener('load', t.step_func(() => {
    const iframeTimeline = iframe.contentDocument.timeline;
    const valueAtStart   = iframeTimeline.currentTime;
    const timeAtStart    = window.performance.now();
    while (iframe.contentWindow.performance.now() - timeAtStart < 50) {
      // Wait 50ms
    }
    assert_equals(iframeTimeline.currentTime, valueAtStart,
      'iframe document.timeline.currentTime does not change within a '
      + ' script block');

    iframe.contentWindow.requestAnimationFrame(t.step_func_done(() => {
      assert_greater_than(iframeTimeline.currentTime, valueAtStart,
        'iframe document.timeline.currentTime increases between script blocks');
      iframe.remove();
    }));
  }));

  document.body.appendChild(iframe);
}, 'iframe document.timeline.currentTime liveness tests');

async_test(function(t) {
  const startTime = document.timeline.currentTime;
  let firstRafTime;

  requestAnimationFrame(function() {
    t.step(function() {
      assert_greater_than_equal(document.timeline.currentTime, startTime,
                                'currentTime should have progressed');
      firstRafTime = document.timeline.currentTime;
    });
  });

  requestAnimationFrame(function() {
    t.step(function() {
      assert_equals(document.timeline.currentTime, firstRafTime,
                    'currentTime should be the same');
    });
    t.done();
  });
}, 'document.timeline.currentTime time should be the same for all RAF'
   + ' callbacks in a frame');

</script>
